Add files via upload
Trained YOLOv9 model to detect potatoes in 5 categories. 

Hyperparameter Changes:
○ epochs=5 → 1
■ Reducing the number of epochs decreased the total number of
times the entire dataset was passed forward and backward through
the neural network. This cut down the overall training time
dramatically, from potentially 15-20 hours to just 3-4 hours. This
also caused underfitting and provided a more practical solution for
showcasing capabilities for further development
○ batch=8 → 30
■ Increasing the batch size allowed more data to be processed
simultaneously during each iteration of an epoch. This not only
speeds up the training process due to fewer updates needed to the
model's weights but also helps in stabilizing the gradient estimates,
potentially enhancing the learning process. This required a lot of
ram, but the decrease in image size compensated for this, but
compromised the overall clarity of detections.
○ image size=640 → 416
■ Reducing the input image size decreased the number of pixels the
model had to process, lowering the computational load and
speeding up training. This caused the model to generalize on
tracking from the YouTube video as it didn’t learn the fine details
and kept tracking round objects. Overall best.pt had less spatial
information to learn from, which is usually a trade-off between
speed and accuracy.
○ Learning rate=0.0001 → 0.001
■ This was to ensure that each epoch made sufficient progress,
which updated model weights more significantly with each batch
when the total number of epochs was limited to just one.